+++ strangemood/programs/strangemood/src/util.rs
@@ -1,0 +1,3 @@
+pub fn amount_as_float(amount: u64, decimals: u8) -> f64 {
+    amount as f64 / i32::pow(10, decimals.into()) as f64
+}
